+++ packages/vue-datepicker/vite.config.ts
@@ -46,6 +46,7 @@ export default defineConfig({
     },
     server: {
         port: 3000,
+        host: true,
         open: true,
         cors: true,
         allowedHosts: [